Order properties in the expected order
Extract read-only subtrait of AbstractGraphSource
