+++ spark-cypher-testing/src/test/scala/org/opencypher/spark/api/io/neo4j/Neo4jReadOnlySourceUnitTest.scala
@@ -1,10 +1,36 @@
+/*
+ * Copyright (c) 2016-2018 "Neo4j Sweden, AB" [https://neo4j.com]
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Attribution Notice under the terms of the Apache License 2.0
+ *
+ * This work was created by the collective efforts of the openCypher community.
+ * Without limiting the terms of Section 6, any Derivative Work that is not
+ * approved by the public consensus process of the openCypher Implementers Group
+ * should not be described as “Cypher” (and Cypher® is a registered trademark of
+ * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
+ * proposals for change that have been documented or implemented should only be
+ * described as "implementation extensions to Cypher" or as "proposed changes to
+ * Cypher that are not yet approved by the openCypher community".
+ */
 package org.opencypher.spark.api.io.neo4j
 
 import java.net.URI
 
 import org.opencypher.okapi.api.graph.GraphName
 import org.opencypher.okapi.api.schema.Schema
-import org.opencypher.okapi.api.types.{CTFloat, CTInteger, CTString}
+import org.opencypher.okapi.api.types.{CTBoolean, CTFloat, CTInteger, CTString}
 import org.opencypher.okapi.testing.BaseTestSuite
 import org.opencypher.spark.api.CAPSSession
 
@@ -13,7 +39,7 @@ class Neo4jReadOnlySourceUnitTest extends BaseTestSuite {
   private val schema = Schema.empty
     .withNodePropertyKeys("A")("foo" -> CTInteger, "bar" -> CTString.nullable)
     .withNodePropertyKeys("B")()
-    .withRelationshipPropertyKeys("TYPE")("foo" -> CTFloat.nullable)
+    .withRelationshipPropertyKeys("TYPE")("foo" -> CTFloat.nullable, "f" -> CTBoolean)
     .withRelationshipPropertyKeys("TYPE2")()
 
   private val entireGraph = "allOfIt"
@@ -21,7 +47,7 @@ class Neo4jReadOnlySourceUnitTest extends BaseTestSuite {
 
   it("constructs flat node queries from schema") {
     pgds.flatNodeQuery(GraphName(entireGraph), Set("A"), schema) should equal(
-      "MATCH (n:A) RETURN id(n) AS id, n.foo, n.bar"
+      "MATCH (n:A) RETURN id(n) AS id, n.bar, n.foo"
     )
   }
 
@@ -33,7 +59,7 @@ class Neo4jReadOnlySourceUnitTest extends BaseTestSuite {
 
   it("constructs flat relationship queries from schema") {
     pgds.flatRelQuery(GraphName(entireGraph), "TYPE", schema) should equal(
-      "MATCH ()-[r:TYPE]->() RETURN id(r) AS id, r.foo"
+      "MATCH ()-[r:TYPE]->() RETURN id(r) AS id, r.f, r.foo"
     )
   }
